fix: 1.21.1 API alignment (InteractionResultHolder, ChunkStatus pkg, ticker, seed via RandomState, int ambient interval)

+++ src/main/java/net/backrooms/entity/BacteriaEntity.java
@@ -78,7 +78,7 @@ public class BacteriaEntity extends Monster {
 	}
 
 	@Override
-	public float getAmbientSoundInterval() {
-		return 60.0F;
+	public int getAmbientSoundInterval() {
+		return 60;
 	}
 }
